api/requirements.txt
django==2.0.3
djangorestframework==3.7.7
django-environ==0.4.3
psycopg2==2.6.2
environ==1.0
requests==2.18.4